return org spaces' running status
